move discipline filter into a popover #243
